Fix intended rotten behaviour for comestibles

+++ src/eat.c
@@ -2241,7 +2241,7 @@ doeat()		/* generic "eat" command funtion (see cmd.c) */
 	    victual.reqtime = objects[otmp->otyp].oc_delay;
 	    if (otmp->otyp != FORTUNE_COOKIE &&
 		(otmp->cursed ||
-		 (((monstermoves - otmp->age) > (int) otmp->blessed ? 50:30) &&
+		 (((monstermoves - otmp->age) > (otmp->blessed ? 50 : 30)) &&
 		(otmp->orotten || !rn2(7))))) {
 
 		if (rottenfood(otmp)) {
